impl clock attack
